ref cascade

+++ ge/single_op/single_op_model.cc
@@ -613,7 +613,7 @@ Status SingleOpModel::BuildDynamicOp(StreamResource &resource, DynamicSingleOp &
   single_op.num_inputs_ = data_ops_.size();
   single_op.num_outputs_ = netoutput_op_->GetAllInputsSize();
   GE_CHK_STATUS_RET_NOLOG(InitModelMem(resource));
-  model_params_.memory_size = UINT_MAX;
+  model_params_.memory_size = UINT64_MAX;
   model_params_.graph_is_dynamic = true;
 
   auto ge_model = model_helper_.GetGeModel();
